# cmd/heimdall-token-analytics/api/swagger.yaml
# SPDX-License-Identifier: ice License 1.0

definitions:
  main.ExternalDataRequestBody:
    properties:
      creatorAvatar:
        example: https://example.com/avatar.png
        type: string
      creatorDisplayName:
        example: John Doe
        type: string
      creatorUsername:
        example: johndoe
        type: string
      creatorVerified:
        example: true
        type: boolean
    type: object
  main.SessionViewCreateResponse:
    properties:
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      ttl:
        description: Session TTL in seconds
        example: 1800
        type: integer
    type: object
  server.ResponseErrorBody:
    properties:
      code:
        example: SOMETHING_NOT_FOUND
        type: string
      error:
        example: something is missing
        type: string
    type: object
  tokenanalytics.Addresses:
    properties:
      ionConnect:
        type: string
      twitter:
        type: string
    type: object
  tokenanalytics.BondingCurveProgress:
    properties:
      currentAmount:
        type: number
      currentAmountUSD:
        type: number
      goalAmount:
        type: number
      goalAmountUSD:
        type: number
    type: object
  tokenanalytics.CommunityToken:
    properties:
      addresses:
        $ref: '#/definitions/tokenanalytics.Addresses'
      createdAt:
        type: string
      creator:
        $ref: '#/definitions/tokenanalytics.User'
      description:
        type: string
      imageUrl:
        type: string
      marketData:
        $ref: '#/definitions/tokenanalytics.MarketData'
      title:
        type: string
      type:
        type: string
    type: object
  tokenanalytics.HolderPosition:
    properties:
      amount:
        type: integer
      amountUSD:
        type: number
      holder:
        $ref: '#/definitions/tokenanalytics.User'
      rank:
        type: integer
      supplyShare:
        type: number
    type: object
  tokenanalytics.MarketData:
    properties:
      bondingCurveProgress:
        $ref: '#/definitions/tokenanalytics.BondingCurveProgress'
      holders:
        type: integer
      marketCap:
        type: number
      position:
        $ref: '#/definitions/tokenanalytics.Position'
      priceUSD:
        type: number
      ticker:
        type: string
      topHolders:
        items:
          $ref: '#/definitions/tokenanalytics.HolderPosition'
        type: array
      volume:
        type: number
    type: object
  tokenanalytics.OHLCV:
    properties:
      close:
        type: number
      high:
        type: number
      low:
        type: number
      open:
        type: number
      timestamp:
        type: integer
      volume:
        type: number
    type: object
  tokenanalytics.Position:
    properties:
      amount:
        type: integer
      amountUSD:
        type: number
      pnl:
        type: number
      pnlPercentage:
        type: number
      rank:
        type: integer
    type: object
  tokenanalytics.TopHolderPosition:
    properties:
      creator:
        $ref: '#/definitions/tokenanalytics.User'
      position:
        $ref: '#/definitions/tokenanalytics.HolderPosition'
    type: object
  tokenanalytics.Trade:
    properties:
      creator:
        $ref: '#/definitions/tokenanalytics.User'
      position:
        $ref: '#/definitions/tokenanalytics.TradePosition'
    type: object
  tokenanalytics.TradePosition:
    properties:
      addresses:
        $ref: '#/definitions/tokenanalytics.Addresses'
      amount:
        type: integer
      amountUSD:
        type: number
      balance:
        type: integer
      balanceUSD:
        type: number
      createdAt:
        type: string
      holder:
        $ref: '#/definitions/tokenanalytics.User'
      type:
        $ref: '#/definitions/tokenanalytics.TradeType'
    type: object
  tokenanalytics.TradeStats:
    properties:
      1h:
        $ref: '#/definitions/tokenanalytics.TradeStatsAggregate'
      5m:
        $ref: '#/definitions/tokenanalytics.TradeStatsAggregate'
      6h:
        $ref: '#/definitions/tokenanalytics.TradeStatsAggregate'
      24h:
        $ref: '#/definitions/tokenanalytics.TradeStatsAggregate'
    type: object
  tokenanalytics.TradeStatsAggregate:
    properties:
      buysTotalAmountUSD:
        type: number
      netBuy:
        type: number
      numberOfBuys:
        type: integer
      numberOfSells:
        type: integer
      sellsTotalAmountUSD:
        type: number
      volumeUSD:
        type: number
    type: object
  tokenanalytics.TradeType:
    enum:
    - buy
    - sell
    type: string
    x-enum-varnames:
    - tradeTypeBuy
    - tradeTypeSell
  tokenanalytics.User:
    properties:
      addresses:
        $ref: '#/definitions/tokenanalytics.Addresses'
      avatar:
        type: string
      display:
        type: string
      name:
        type: string
      verified:
        type: boolean
    type: object
info:
  contact:
    name: ice.io
    url: https://ice.io
  description: |-
    This service provides analytics data for various tokens.
    URLs starting with /v1/sse return Server-Sent Events (SSE) for real-time updates.
    URLs starting with /v1/ws utilize WebSockets for bidirectional communication.
  title: Token Analytics Service API.
  version: latest
paths:
  /v1/community-tokens:
    get:
      description: Returns community tokens information for the given Ion Connect
        addresses.
      parameters:
      - description: External addresses of the tokens
        example: 0x1234...,0x5678...
        in: query
        items:
          type: string
        name: externalAddresses
        required: true
        type: array
      - description: Number of top holders to include (1-10)
        example: 3
        in: query
        maximum: 10
        minimum: 1
        name: includeTopHolders
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/tokenanalytics.CommunityToken'
            type: array
        "401":
          description: if auth token is missing or invalid
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "504":
          description: if request times out
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
      security:
      - Nostr: []
      - XCom: []
      tags:
      - Tokens
  /v1/community-tokens/{externalAddressOrViewType}:
    get:
      description: Returns community tokens information for the given Ion Connect
        addresses.
      parameters:
      - description: View type (latest)
        example: '"latest"'
        in: path
        name: externalAddressOrViewType
        required: true
        type: string
      - description: Search keyword
        example: '"bitcoin"'
        in: query
        name: keyword
        type: string
      - description: Number of items to return
        example: 10
        format: int32
        in: query
        name: limit
        type: integer
      - description: Number of items to skip
        example: 0
        format: int32
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/tokenanalytics.CommunityToken'
            type: array
        "401":
          description: if auth token is missing or invalid
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "504":
          description: if request times out
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
      security:
      - Nostr: []
      - XCom: []
      tags:
      - Tokens
  /v1/community-tokens/{externalAddressOrViewType}/external-data:
    put:
      consumes:
      - application/json
      description: Syncs external creator information for a community token.
      parameters:
      - description: External address
        in: path
        name: externalAddressOrViewType
        required: true
        type: string
      - description: Creator information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.ExternalDataRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK - Data synced successfully
        "400":
          description: if request body is invalid
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "401":
          description: if auth token is missing or invalid
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "504":
          description: if request times out
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
      security:
      - Nostr: []
      - XCom: []
      tags:
      - Tokens
  /v1/community-tokens/{externalAddressOrViewType}/latest-trades:
    get:
      description: Returns trade history for a specific community token address.
      parameters:
      - description: External address
        example: '"0x1234..."'
        in: path
        name: externalAddressOrViewType
        required: true
        type: string
      - description: Number of items to return
        example: 10
        format: int32
        in: query
        name: limit
        type: integer
      - description: Number of items to skip
        example: 0
        format: int32
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/tokenanalytics.Trade'
            type: array
        "401":
          description: if auth token is missing or invalid
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "504":
          description: if request times out
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
      security:
      - Nostr: []
      - XCom: []
      tags:
      - Tokens
  /v1/community-tokens/{externalAddressOrViewType}/viewing-sessions:
    post:
      description: Creates a new session view for community tokens analytics.
      parameters:
      - description: View type
        enum:
        - top
        - trending
        - bondingCurveProgress
        example: '"top"'
        in: path
        name: externalAddressOrViewType
        required: true
        type: string
      - description: Token type filter
        enum:
        - profile
        - post
        - video
        - article
        example: '"profile"'
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.SessionViewCreateResponse'
        "401":
          description: if auth token is missing or invalid
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "504":
          description: if request times out
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
      security:
      - Nostr: []
      - XCom: []
      tags:
      - Tokens
  /v1/community-tokens/{externalAddressOrViewType}/viewing-sessions/{viewingSessionId}:
    get:
      description: Returns community tokens information for a specific viewing session.
      parameters:
      - description: View type
        enum:
        - top
        - trending
        - bondingCurveProgress
        example: '"top"'
        in: path
        name: externalAddressOrViewType
        required: true
        type: string
      - description: Viewing session ID
        example: '"550e8400-e29b-41d4-a716-446655440000"'
        in: path
        name: viewingSessionId
        required: true
        type: string
      - description: Search keyword
        example: '"bitcoin"'
        in: query
        name: keyword
        type: string
      - description: Number of items to return
        example: 10
        format: int32
        in: query
        name: limit
        type: integer
      - description: Number of items to skip
        example: 0
        format: int32
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/tokenanalytics.CommunityToken'
            type: array
        "401":
          description: if auth token is missing or invalid
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "504":
          description: if request times out
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
      security:
      - Nostr: []
      - XCom: []
      tags:
      - Tokens
  /v1sse/community-tokens:
    get:
      description: Streams community tokens information for the given Ion Connect
        addresses.
      parameters:
      - description: External addresses of the tokens
        example: 0x1234...,0x5678...
        in: query
        items:
          type: string
        name: externalAddresses
        required: true
        type: array
      - description: Number of top holders to include (1-10)
        example: 3
        in: query
        maximum: 10
        minimum: 1
        name: includeTopHolders
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tokenanalytics.CommunityToken'
        "401":
          description: if auth token is missing or invalid
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "504":
          description: if request times out
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
      security:
      - Nostr: []
      - XCom: []
      tags:
      - stream
  /v1sse/community-tokens/{externalAddressOrViewType}:
    get:
      description: Streams community tokens information for the given type.
      parameters:
      - description: View type (latest, featured, top, trending, or bondingCurveProgress)
        example: '"latest","featured","top","trending","bondingCurveProgress"'
        in: path
        name: externalAddressOrViewType
        required: true
        type: string
      - description: Viewing session ID (required for top/trending/bondingCurveProgress)
        example: '"550e8400-e29b-41d4-a716-446655440000"'
        in: query
        name: viewingSessionId
        type: string
      - description: Token type filter (profile, post, video, or article)
        example: '"profile"'
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/tokenanalytics.CommunityToken'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "401":
          description: if auth token is missing or invalid
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "504":
          description: if request times out
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
      security:
      - Nostr: []
      - XCom: []
      tags:
      - stream
  /v1sse/community-tokens/{externalAddressOrViewType}/latest-trades:
    get:
      description: Streams latest trades for a specific community token address.
      parameters:
      - description: External address
        example: '"0x1234..."'
        in: path
        name: externalAddressOrViewType
        required: true
        type: string
      - description: Number of items to return
        example: 10
        format: int32
        in: query
        name: limit
        type: integer
      - description: Number of items to skip
        example: 0
        format: int32
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tokenanalytics.Trade'
        "401":
          description: if auth token is missing or invalid
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "504":
          description: if request times out
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
      security:
      - Nostr: []
      - XCom: []
      tags:
      - stream
  /v1sse/community-tokens/{externalAddressOrViewType}/ohlcv:
    get:
      description: Streams OHLCV (Open, High, Low, Close, Volume) data for a specific
        community token address.
      parameters:
      - description: External address
        example: '"0x1234..."'
        in: path
        name: externalAddressOrViewType
        required: true
        type: string
      - description: Time interval
        example: '"1m"'
        in: query
        name: interval
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tokenanalytics.OHLCV'
        "401":
          description: if auth token is missing or invalid
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "504":
          description: if request times out
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
      security:
      - Nostr: []
      - XCom: []
      tags:
      - stream
  /v1sse/community-tokens/{externalAddressOrViewType}/top-holders:
    get:
      description: Streams top holders information for a specific community token
        address.
      parameters:
      - description: External address
        example: '"0x1234..."'
        in: path
        name: externalAddressOrViewType
        required: true
        type: string
      - description: Number of items to return
        example: 10
        format: int32
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/tokenanalytics.TopHolderPosition'
            type: array
        "401":
          description: if auth token is missing or invalid
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "504":
          description: if request times out
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
      security:
      - Nostr: []
      - XCom: []
      tags:
      - stream
  /v1sse/community-tokens/{externalAddressOrViewType}/trading-stats:
    get:
      description: Streams trading statistics for a specific community token address.
      parameters:
      - description: External address
        example: '"0x1234..."'
        in: path
        name: externalAddressOrViewType
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tokenanalytics.TradeStats'
        "401":
          description: if auth token is missing or invalid
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "504":
          description: if request times out
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
      security:
      - Nostr: []
      - XCom: []
      tags:
      - stream
  /v1ws/community-tokens:
    get:
      description: Streams community tokens information for the given Ion Connect
        addresses.
      parameters:
      - description: External addresses of the tokens
        example: 0x1234...,0x5678...
        in: query
        items:
          type: string
        name: externalAddresses
        required: true
        type: array
      - description: Number of top holders to include (1-10)
        example: 3
        in: query
        maximum: 10
        minimum: 1
        name: includeTopHolders
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tokenanalytics.CommunityToken'
        "401":
          description: if auth token is missing or invalid
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "504":
          description: if request times out
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
      security:
      - Nostr: []
      - XCom: []
      tags:
      - stream
  /v1ws/community-tokens/{externalAddressOrViewType}:
    get:
      description: Streams community tokens information for the given type.
      parameters:
      - description: View type (latest, featured, top, trending, or bondingCurveProgress)
        example: '"latest","featured","top","trending","bondingCurveProgress"'
        in: path
        name: externalAddressOrViewType
        required: true
        type: string
      - description: Viewing session ID (required for top/trending/bondingCurveProgress)
        example: '"550e8400-e29b-41d4-a716-446655440000"'
        in: query
        name: viewingSessionId
        type: string
      - description: Token type filter (profile, post, video, or article)
        example: '"profile"'
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/tokenanalytics.CommunityToken'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "401":
          description: if auth token is missing or invalid
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "504":
          description: if request times out
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
      security:
      - Nostr: []
      - XCom: []
      tags:
      - stream
  /v1ws/community-tokens/{externalAddressOrViewType}/latest-trades:
    get:
      description: Streams latest trades for a specific community token address.
      parameters:
      - description: External address
        example: '"0x1234..."'
        in: path
        name: externalAddressOrViewType
        required: true
        type: string
      - description: Number of items to return
        example: 10
        format: int32
        in: query
        name: limit
        type: integer
      - description: Number of items to skip
        example: 0
        format: int32
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tokenanalytics.Trade'
        "401":
          description: if auth token is missing or invalid
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "504":
          description: if request times out
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
      security:
      - Nostr: []
      - XCom: []
      tags:
      - stream
  /v1ws/community-tokens/{externalAddressOrViewType}/ohlcv:
    get:
      description: Streams OHLCV (Open, High, Low, Close, Volume) data for a specific
        community token address.
      parameters:
      - description: External address
        example: '"0x1234..."'
        in: path
        name: externalAddressOrViewType
        required: true
        type: string
      - description: Time interval
        example: '"1m"'
        in: query
        name: interval
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tokenanalytics.OHLCV'
        "401":
          description: if auth token is missing or invalid
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "504":
          description: if request times out
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
      security:
      - Nostr: []
      - XCom: []
      tags:
      - stream
  /v1ws/community-tokens/{externalAddressOrViewType}/top-holders:
    get:
      description: Streams top holders information for a specific community token
        address.
      parameters:
      - description: External address
        example: '"0x1234..."'
        in: path
        name: externalAddressOrViewType
        required: true
        type: string
      - description: Number of items to return
        example: 10
        format: int32
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/tokenanalytics.TopHolderPosition'
            type: array
        "401":
          description: if auth token is missing or invalid
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "504":
          description: if request times out
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
      security:
      - Nostr: []
      - XCom: []
      tags:
      - stream
  /v1ws/community-tokens/{externalAddressOrViewType}/trading-stats:
    get:
      description: Streams trading statistics for a specific community token address.
      parameters:
      - description: External address
        example: '"0x1234..."'
        in: path
        name: externalAddressOrViewType
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tokenanalytics.TradeStats'
        "401":
          description: if auth token is missing or invalid
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
        "504":
          description: if request times out
          schema:
            $ref: '#/definitions/server.ResponseErrorBody'
      security:
      - Nostr: []
      - XCom: []
      tags:
      - stream
schemes:
- https
securityDefinitions:
  Nostr:
    description: Type "Nostr" followed by a space and your Nostr NIP-42 compatible
      base64-encoded event.
    in: header
    name: Authorization
    type: apiKey
  XCom:
    description: Type "X.com" followed by a space and your X.com JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"